patito/requirements.txt
antlr4-python3-runtime==4.13.2
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
